Windows 10
I am removing Windows 7 and 8.1 compatibility. Windows 10 is the future
for Microsoft, and it is now being used by businesses and home users.

TRON proper works for the old systems. I want this version to be
optimized for Windows 10, using some of the new features that it offers.
I will be using some of the Powershell Gallery components to supplement
TRON:Evo, and breaking backwards compatibility in the process.

Eventually, this script can be run on **ANY** Windows 10-based system,
from Win Server 2016 (support coming upon release), Home, Pro,
Enterprise, IOT, and XBOne.

These are the initial changes moving towards 2.0 of Tron:EVO.

Also changing from PowerGUI to VSCode for the development environment.
The .JSON files are now part of the repository, but not the release
product.